2nd part of the tasks

+++ additional_tasks/find_sum_numbers/src/utils.hpp
@@ -1,5 +1,6 @@
 #pragma once
-
+#include <stdexcept>
 #include <vector>
 
-std::pair<int, int> GetNumbersOfSum(const std::vector<int>& numbers, int sum);
+std::pair<size_t, size_t> GetNumbersOfSum(const std::vector<int>& numbers,
+                                          int sum);
